[python] Extract string handling into dedicated string_handler class
Moved 26 string manipulation methods (~1,500 lines) from python_converter
into a new string_handler class to improve code organization and
maintainability.

Key changes:
- Created string_handler class with all string operations (size, conversion,
  f-strings, concatenation, comparison, and string methods)
- Updated python_converter to delegate string operations via string_handler_
- Modified string_builder to use string_handler* for array utilities
- Added accessor method get_string_handler() for cross-class access
- Updated function_call_builder and python_list to use new accessor

This refactoring reduces python_converter complexity by 26% while preserving
all functionality. No behavioral changes.
